Web.Larceny

Description

A long description with example use with Scotty.

It'll be something like:

Write some templates.

Add a Library to your Scotty state with loadTemplates.

Write some Substitutions.

Write some Larceny/Scotty glue.

Use render and renderWith in your Scotty handlers.

Admire your lovely app!!

Documentation

newtype Blank

A "blank" in the template that can be filled in with some value by Haskell code. Blanks can be tags or they can be all or parts of attribute values in tags.

Example blanks:

<skater>                           <- "skater"
<p class=${name}>                  <- "name"
<skater name="${name}">            <- both "skater" and "name"
<a href="teams/${team}/{$number}"> <- both "team" and number"

Constructors

Blank Text

newtype Fill s

A Fill is how to fill in a Blank.

In most cases, you can use helper functions like textFill or fillChildrenWith to create your fills. You can also write Fills from scratch.

Fill $ attrs _tpl _lib ->
         return $ T.pack $ show $ M.keys attrs)

With that Fill, a Blank like this:

<displayAttrs attribute="hello!" another="goodbye!"/>

would be rendered as:

["attribute", "another"]

Fills (and Substitutions and Templates) have the type `StateT s IO Text` in case you need templates to depend on IO actions (like looking something up in a database) or store state (perhaps keeping track of what's already been rendered).

Constructors

Fill

Fields

unFill :: Attributes -> (Path, Template s) -> Library s -> StateT s IO Text

type Attributes = Map Text Text

The Blank's attributes, a map from the attribute name to it's value.

type Substitutions s = Map Blank (Fill s)

A map from Blanks to how to fill in the Blank.

newtype Template s

When you run a Template with the path, some substitutions, and the template library, you'll get back some stateful text.

Use loadTemplates to load the templates from some directory into a template library. Use the render functions to render templates by path.

Constructors

Template

Fields

runTemplate :: Path -> Substitutions s -> Library s -> StateT s IO Text

type Path = [Text]

The path to a template.

type Library s = Map Path (Template s)

A collection of templates.

data Overrides

If no substitutions are given, Larceny only understands valid HTML 5 tags. It won't attempt to "fill in" tags that are already valid HTML 5. Use Overrides to use non-HTML 5 tags without providing your own substitutions, or to provide fills for standard HTML tags.

-- Use the deprecated "marquee" and "blink" tags and write your
-- own fill for the "a" tag.
Overrides ["marquee", "blink"] ["a"]

Constructors

Overrides

Fields

customPlainNodes :: [Text]

overrideNodes :: [Text]

defaultOverrides :: Overrides

Default uses no overrides.

render :: Library s -> s -> Path -> IO (Maybe Text)

Render a template from the library by path.

render appTemplates appState ["path", "to", "template"]

renderWith :: Library s -> Substitutions s -> s -> Path -> IO (Maybe Text)

Render a template from the library by path, with some additional substitutions.

renderWith appTemplates extraSubs appState ["path", "to", "template"]

renderRelative :: Library s -> Substitutions s -> s -> Path -> Path -> IO (Maybe Text)

Render a template found relative to current template's path.

This will attempt to find the target template starting at the same level as the given path, then will traverse up the directory tree until it finds a template with the target path.

For example: Given these templates: ["current"], ["current", "dashboard"], ["current", "private", "dashboard"], ["private", "dashboard"], renderRelative called with a given path of ["current"] and target path of ["private", "dashboard"] will find ["current", "private", "dashboard"]. If there wasn't a ["current", "private", "dashboard"], it would render ["private", "dashboard"].

loadTemplates :: FilePath -> Overrides -> IO (Library s)

Load all the templates in some directory into a Library.

subs :: [(Text, Fill s)] -> Substitutions s

Turn tuples of text and fills to Substitutions.

subs [("blank", textFill "the fill")
     ,("another-blank", textFill "another fill")]

textFill :: Text -> Fill s

A plain text fill.

textFill "This text will be displayed in place of the blank"

textFill' :: StateT s IO Text -> Fill s

Use state or IO, then fill in some text.

-- getTextFromDatabase :: StateT () IO Text
textFill' getTextFromDatabase

mapSubs :: (a -> Substitutions s) -> [a] -> Fill s

Create substitutions for each element in a list and fill the child nodes with those substitutions.

<members><name /></members>
("members", mapSubs (name -> subs [("name", textFill name)])
                    ["Bonnie Thunders", "Donna Matrix", "Beyonslay"]

Bonnie Thunders Donna Matrix Beyonslay

mapSubs' :: (a -> StateT s IO (Substitutions s)) -> [a] -> Fill s

Create substitutions for each element in a list (using IO/state if needed) and fill the child nodes with those substitutions.

fillChildren :: Fill s

Fill in the child nodes of the blank with substitutions already available.

<no-op><p>Same</p></no-op>
("no-op", fillChildren)

<p>Same</p>

fillChildrenWith :: Substitutions s -> Fill s

Fill in the child nodes of the blank with new substitutions.

<member><name /></member>
("skater", fillChildrenWith (subs $ [("name", textFill "Bonnie Thunders")]))

Beyonslay

fillChildrenWith' :: StateT s IO (Substitutions s) -> Fill s

Use substitutions with State and IO.

<changeTheWorld><results /></changeTheWorld>
-- doABunchOfStuffAndGetSubstitutions :: StateT () IO (Substitutions ())
("changeTheWorld", fillChildrenWith' doStuffAndGetSubstitutions)

This template did IO!

maybeFillChildrenWith :: Maybe (Substitutions s) -> Fill s

Fill with substitutions if those substitutions are provided.

<ifDisplayUser><userName /></ifDisplayUser>
("ifDisplayUser", maybeFillChildrenWith
                    (Just $ subs' ("userName", textFill "Bonnie Thunders")))

Bonnie Thunders

maybeFillChildrenWith' :: StateT s IO (Maybe (Substitutions s)) -> Fill s

Use state and IO and maybe fill in with some substitutions.

<ifLoggedIn>Logged in as <userName /></ifLoggedIn>
("ifLoggedIn", maybeFillChildrenWith' $ do
                 mUser <- getLoggedInUser -- returns (Just "Bonnie Thunders")
                 case mUser of
                   Just user -> Just $ subs' ("userName", textFill user)
                   Nothing   -> Nothing)

Bonnie Thunders

useAttrs :: (Attributes -> k -> Fill s) -> k -> Fill s

Use attributes from the the blank as arguments to the fill.

<desc length="10" />
("desc", useAttrs (a"length") descriptionFill)
descriptionFill len = textFill $ T.take len
                                 "A really long description"
                                 <> "..."))

A really l...

useAttrs takes two arguments. The first is a way to get values of attributes that you can use in Fills. You can use a and % to create these. The second argument is a function that uses the values of those attributes to create a Fill.

class FromAttribute a where

A typeclass for things that can be parsed from attributes.

Methods

fromAttribute :: Maybe Text -> Either AttrError a

data AttrError

If an attribute is required but missing, or unparsable, one of these errors is thrown.

Constructors

AttrMissing

AttrUnparsable Text

AttrOtherError Text

a :: FromAttribute a => Text -> Attributes -> (a -> b) -> b

Prepend a to the name of an attribute to pass the value of that attribute to the fill.

The type of the attribute is whatever type the fill expects. If a can't parse the value, then there will be an error when the template is rendered.

(%) :: (Attributes -> a -> b) -> (Attributes -> b -> c) -> Attributes -> a -> c

Use with a to use multiple attributes in the fill.

<desc length="10" />
("desc", useAttrs (a"length" % a"ending") descriptionFill)
descriptionFill len maybeEnding =
  let ending = fromMaybe "..." maybeEnding in
  textFill $ T.take n
             "A really long description"
             <> ending))

A really l...

parse :: Text -> Template s

Turn lazy text into templates.

parseWithOverrides :: Overrides -> Text -> Template s

Use overrides when parsing a template.